Delete settings.route-lazyload, move its content to settings.route
